+++ srcs/main_flow/init.c
@@ -6,7 +6,7 @@
 /*   By: yioffe <yioffe@student.42lisboa.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/24 11:14:33 by yioffe            #+#    #+#             */
-/*   Updated: 2024/05/29 17:37:24 by yioffe           ###   ########.fr       */
+/*   Updated: 2024/05/30 17:08:24 by yioffe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,8 @@ void	new_env(t_shell *shell)
 {
 	char	*default_env[4];
 
-	default_env[0] = "PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin";
+	default_env[0] = \
+		"PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin";
 	default_env[1] = "PWD=/";
 	default_env[2] = "USER=evaluator";
 	default_env[3] = NULL;
@@ -133,17 +134,3 @@ int	init_shell(t_shell *shell, char **env)
 	shell->exit_status = EXIT_SUCCESS;
 	return (EXIT_SUCCESS);
 }
-
-// if (getenv("USER") == NULL)// ASK YULIA
-// 	init_missing_environment(shell, env);
-// else
-// {
-// 	shell->envless = FALSE;
-// 	shell->color_codes = TRUE;
-// 	shell->env_head = init_env(env);
-// 	shell->user_name = getenv("USER");
-// 	shell->cmd_paths = ft_split(get_path(env), ':');
-// }https://github.com/zstenger93/minishell/blob/master/src/init/init.c#L15
-
-// Replace the environment variables into a linked list so you can delete or add to them later using export and unset builtins. In addition to displaying them using env or export (without arguments) builtins.
-// https://github.com/fraqioui/minishell?tab=readme-ov-file#Initializing
